new+1

+++ blitz3d/texture.cpp
@@ -21,7 +21,7 @@ static vector<Filter> filters;
 static int filterFile( const string &t,int flags ){
 	//check filters...
 	string l=tolower(t);
-	for( int k=0;k<(int)filters.size();++k ){
+	for( int k=0;k<filters.size();++k ){
 		if( l.find( filters[k].t )!=string::npos ){
 			flags|=filters[k].flags;
 		}
@@ -37,6 +37,7 @@ struct Texture::Rep{
 
 	int tex_blend,tex_flags;
 	bool transparent;
+
 	DWORD bumpEnvMat[2][2];
 	DWORD bumpEnvScale;
 	DWORD bumpEnvOffset;
@@ -145,18 +146,18 @@ void Texture::setFlags( int flags ){
 	rep->tex_flags=flags;
 }
 
-void Texture::setBumpEnvMat( int x,int y,float envmat ){
-	if( !rep ) return;
+void Texture::setBumpEnvMat(int x, int y, float envmat) {
+	if (!rep) return;
 	rep->bumpEnvMat[x][y] = *((DWORD*)&envmat);
 }
 
-void Texture::setBumpEnvScale( float envscale ){
-	if( !rep ) return;
+void Texture::setBumpEnvScale(float envscale) {
+	if (!rep) return;
 	rep->bumpEnvScale = *((DWORD*)&envscale);
 }
 
-void Texture::setBumpEnvOffset( float envoffset ){
-	if( !rep ) return;
+void Texture::setBumpEnvOffset(float envoffset) {
+	if (!rep) return;
 	rep->bumpEnvOffset = *((DWORD*)&envoffset);
 }
 
@@ -165,7 +166,7 @@ bool Texture::isTransparent()const{
 }
 
 gxCanvas *Texture::getCanvas( int n )const{
-	return rep && n>=0 && n<(int)(rep->tex_frames.size()) ? rep->tex_frames[n] : 0;
+	return rep && n>=0 && n<rep->tex_frames.size() ? rep->tex_frames[n] : 0;
 }
 
 int Texture::getCanvasFlags()const{
@@ -184,15 +185,15 @@ int Texture::getFlags()const{
 	return rep ? rep->tex_flags : 0;
 }
 
-DWORD Texture::getBumpEnvMat( int x,int y )const{
+DWORD Texture::getBumpEnvMat(int x, int y)const {
 	return rep ? rep->bumpEnvMat[x][y] : 0;
 }
 
-DWORD Texture::getBumpEnvScale()const{
+DWORD Texture::getBumpEnvScale()const {
 	return rep ? rep->bumpEnvScale : 0;
 }
 
-DWORD Texture::getBumpEnvOffset()const{
+DWORD Texture::getBumpEnvOffset()const {
 	return rep ? rep->bumpEnvOffset : 0;
 }
